+++ CHANGELOG.md
@@ -30,6 +30,7 @@
   plugins ([#239][239]).
 * Documentation for the `cpu` module; `interval` has been renamed to
   `poll-interval` ([#241][241]).
+* battery: module was not thread safe.
 
 [239]: https://codeberg.org/dnkl/yambar/issues/239
 [241]: https://codeberg.org/dnkl/yambar/issues/241